+++ Si.GameEngine/Core/TickControllers/StarsSpriteTickController.cs
@@ -15,7 +15,7 @@ namespace Si.GameEngine.Core.TickControllers
         {
         }
 
-        public override void ExecuteWorldClockTick(double epoch, SiVector displacementVector)
+        public override void ExecuteWorldClockTick(float epoch, SiVector displacementVector)
         {
             if (Math.Abs(displacementVector.X) > 1 || Math.Abs(displacementVector.Y) > 1)
             {
